# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservation System API",
        "description": "API REST para el sistema de reservas de hotel. Incluye gestión de amenities, tipos de habitación y habitaciones físicas.",
        "contact": {
            "name": "API Support",
            "email": "adiazy@uamv.edu.ni"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rate-plans": {
            "get": {
                "tags": [
                    "Rate Plans"
                ],
                "summary": "Obtener lista de planes de tarifa",
                "description": "Lista de planes de tarifa\n\nFiltro opcional vía query param:\n- ?is_active=1",
                "operationId": "515679ece11c76fbeaac23eebf272656",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por estado activo (1 o 0)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de planes de tarifa obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rate Plans"
                ],
                "summary": "Crear un nuevo plan de tarifa",
                "description": "Crear un nuevo plan de tarifa",
                "operationId": "e6c5437ea788cf313bc89a4b5478db86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "meal_plan"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "example": "BAR"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "Best Available Rate"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tarifa estándar flexible.",
                                        "nullable": true
                                    },
                                    "cancellation_policy": {
                                        "type": "string",
                                        "example": "Cancelación gratuita hasta 24 horas antes del check-in.",
                                        "nullable": true
                                    },
                                    "meal_plan": {
                                        "description": "RO=Room Only, BB=Bed & Breakfast, HB=Half Board, FB=Full Board, AI=All Inclusive",
                                        "type": "string",
                                        "enum": [
                                            "RO",
                                            "BB",
                                            "HB",
                                            "FB",
                                            "AI"
                                        ],
                                        "example": "RO"
                                    },
                                    "is_refundable": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan de tarifa creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rate-plans/{id}": {
            "get": {
                "tags": [
                    "Rate Plans"
                ],
                "summary": "Obtener un plan de tarifa por ID",
                "description": "Obtener un plan de tarifa específico",
                "operationId": "48e0f348fa60b3ee73b471ab29d9c5d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plan de tarifa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan de tarifa obtenido exitosamente"
                    },
                    "404": {
                        "description": "Plan de tarifa no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rate Plans"
                ],
                "summary": "Actualizar un plan de tarifa",
                "description": "Actualizar un plan de tarifa",
                "operationId": "34a80ebc42795312254c67f56a738c12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plan de tarifa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "meal_plan"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "maxLength": 32
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "cancellation_policy": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "meal_plan": {
                                        "type": "string",
                                        "enum": [
                                            "RO",
                                            "BB",
                                            "HB",
                                            "FB",
                                            "AI"
                                        ]
                                    },
                                    "is_refundable": {
                                        "type": "boolean"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan de tarifa actualizado exitosamente"
                    },
                    "404": {
                        "description": "Plan de tarifa no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rate Plans"
                ],
                "summary": "Desactivar un plan de tarifa",
                "description": "Desactivar un plan de tarifa\n\nOpción B (más realista): En lugar de eliminar físicamente, marcamos is_active = false.\nEsto permite mantener el historial y referencias en reservas existentes.\nAlternativa (Opción A): Podrías usar $ratePlan->delete() para eliminación física.",
                "operationId": "1c292402eaa6bb268d14498da2984a09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plan de tarifa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan de tarifa desactivado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rate plan deactivated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plan de tarifa no encontrado"
                    }
                }
            }
        },
...
    },
...
}